refactor(agent): refine recommendations to be service-type specific
Improves the precision of the recommendations generated in the "partially justified" path.

The logic now checks the `invoice_service` type before providing certain recommendations. For example, it will only suggest benchmarking against other *express* carriers if the invoice itself is for an express service. This makes the agent's advice more contextually relevant and actionable for the user.